feat: only build cpu version in CI

+++ script/post.js
@@ -1,6 +1,7 @@
 const shelljs = require("shelljs")
 const fs = require("node:fs")
 const path = require("node:path")
+const { isCI } = require("std-env")
 
 function cpDirSync(source, target) {
   const stat = fs.statSync(source)
@@ -68,28 +69,29 @@ if(
     }
   }
 
-  // 使用 fs 删除 build 文件夹
-  shelljs.rm("-rf", path.resolve(__dirname, "../lib/whisper.cpp/build"))
-
-  // gpu
-  shelljs.exec("cd ./lib/whisper.cpp && npx cmake-js compile --CDWHISPER_CUBLAS=1 -T whisper-addon -B Release")
-  if(
-    fs.existsSync(path.resolve(__dirname, "../lib/whisper.cpp/build/Release/whisper-addon.node")) 
-    || fs.existsSync(path.resolve(__dirname, "../lib/whisper.cpp/build/bin/Release/whisper-addon.node")) 
-  ) {
-    info("Copy whisper.cpp addon")
-    if(process.platform === "win32") {
-      cpDirSync(
-        path.resolve(__dirname, "../lib/whisper.cpp/build/bin/Release"), 
-        path.resolve(__dirname, "../public/resources/cublas-whisper"),
-      )
-    } else {
-      cpDirSync(
-        path.resolve(__dirname, "../lib/whisper.cpp/build/Release"), 
-        path.resolve(__dirname, "../public/resources/cublas-whisper"),
-      )
-    }
-  } 
+  if(!isCI) {
+    // 使用 fs 删除 build 文件夹
+    shelljs.rm("-rf", path.resolve(__dirname, "../lib/whisper.cpp/build"))
+    // gpu
+    shelljs.exec("cd ./lib/whisper.cpp && npx cmake-js compile --CDWHISPER_CUBLAS=1 -T whisper-addon -B Release")
+    if(
+      fs.existsSync(path.resolve(__dirname, "../lib/whisper.cpp/build/Release/whisper-addon.node")) 
+      || fs.existsSync(path.resolve(__dirname, "../lib/whisper.cpp/build/bin/Release/whisper-addon.node")) 
+    ) {
+      info("Copy whisper.cpp addon")
+      if(process.platform === "win32") {
+        cpDirSync(
+          path.resolve(__dirname, "../lib/whisper.cpp/build/bin/Release"), 
+          path.resolve(__dirname, "../public/resources/cublas-whisper"),
+        )
+      } else {
+        cpDirSync(
+          path.resolve(__dirname, "../lib/whisper.cpp/build/Release"), 
+          path.resolve(__dirname, "../public/resources/cublas-whisper"),
+        )
+      }
+    } 
+  }
 }
 
 if(!fs.existsSync(path.resolve(__dirname, "../lib/vad/CMakeLists.txt"))) {
